+++ Diversity/src/backend/GetAuthors.java
@@ -96,7 +96,7 @@ public class GetAuthors {
 
 	private void dbconnect() {
 		try {
-			cnlocal = dbc.connlocal();
+			cnlocal = Settings.connlocal();
 		} catch (ClassNotFoundException e) {
 			// TODO Auto-generated catch block
 			e.printStackTrace();