+++ Diversity/src/backend/GetInfGraph.java
@@ -61,7 +61,7 @@ public class GetInfGraph {
 
 	private void dbconnect(){
 		try {
-			cnlocal = dbc.connlocal();
+			cnlocal = Settings.connlocal();
 		} catch (ClassNotFoundException e) {
 			// TODO Auto-generated catch block
 			e.printStackTrace();