+++ Diversity/src/backend/GetLastPost.java
@@ -101,7 +101,7 @@ public class GetLastPost {
 
 	private void dbconnect() {
 		try {
-			cnlocal = dbc.connlocal();
+			cnlocal = Settings.connlocal();
 		} catch (ClassNotFoundException e) {
 			// TODO Auto-generated catch block
 			e.printStackTrace();