+++ Diversity/src/backend/GetPopulation.java
@@ -121,7 +121,7 @@ public class GetPopulation {
 
 	private void dbconnect() {
 		try {
-			cnlocal = dbc.connlocal();
+			cnlocal = Settings.connlocal();
 		} catch (ClassNotFoundException e) {
 			// TODO Auto-generated catch block
 			e.printStackTrace();