+++ Diversity/src/backend/GetPosts.java
@@ -179,7 +179,7 @@ public class GetPosts {
 
 	private void dbconnect() {
 		try {
-			cnlocal = dbc.connlocal();
+			cnlocal = Settings.connlocal();
 		} catch (ClassNotFoundException e) {
 			// TODO Auto-generated catch block
 			e.printStackTrace();